gitignore

+++ src/main/java/com/example/ice/Datamon.java
@@ -2,12 +2,12 @@ package com.example.ice;
 
 public interface Datamon {
     public String getName();
-    public int getHP();
+    public double getHP();
     public void setHP(int hp);
-    public int getBaseDamage();
+    public double getBaseDamage();
     public String getCreatureType();
-    public int Damagebuff(Datamon target);
-    public  int takedamage(int damage);
+    public double Damagebuff(Datamon target);
+    public double takedamage(double damage);
     public String introduction();
